daemon: clean up tracker
